Pearl.GaDtTLHT.Properties

Description

Section 1

p00

p00 :: property (h is a list homomorphism)


   (p00)      h (xs ++ ys) = h xs ⊙ h ys

which says that to compute h:

one may arbitrarily split the input list into xs ++ ys,

compute h on them recursively in parallel,

and combine the results using an associative operator ⊙.

Section 2

p01

p01 :: [a1] -> property (e is a unit for a fold)

As is well known, in the world of sets and total functions, the equations:


   (p01)       h  [    ]  =   e
   (p01)       h  (x:xs)  =       (<||) (x, h xs)
   (p01a)      h          = foldr (<||) e

have (p01a) as unique solution for h :: [a] -> b.

p01a

p01a :: property foldr with (a leftwards function) and (a unit e) is unique


   (p01a)      h          = foldr (<||) e

p01b

p01b :: property p01 [pointfree style]

Thus (p01):


   (p01)       h  [    ]  =   e
   (p01)       h  (x:xs)  = (<||) (x, h xs)

can be written:


   (p01b)     h . cons = (<||) . (id >< h)

p02

p02 :: property (foldrr is foldr [resumed])

That is, if:


   (p01a)      h          = foldr (<||) e

then one can show:


   (p02) h (xs ++ ys)   = foldrr (<||) ( xs , h ys )

by induction on xs.

p03

p03 :: property p02 [pointfree style]

Let:

 cat (xs,ys) = xs ++ ys

then equation (p02):

(p02) h (xs ++ ys) = foldrr (<||) ( xs , h ys )

can be point-free as:

(p03) h . cat      = foldrr (<||) . ( id >< h )

p04

p04 :: property (dual to p03) [pointfree]

We have, if h = foldl (||>) e , that:

 (p04)  h . cat = foldlr (||>) . ( h >< id )

Section 3

p05

p05 :: property


   (p05)       splits . k = map (id >< k) . unfoldrp (||>)

p06

p06 :: property


   (p06)       splits . k = map (k >< id) . unfoldlp (<||)

p07

p07 :: property

The equality (p05) can be expressed relationally.

If k = unfoldr (||>) p we have:


  (p07)     cat° . k = (id >< k ) . unfr (||>) p

p08

p08 :: property

For k = unfoldl (<||) p, a symmetric property holds:


  (p08)     cat° . k = (k  >< id) . unfl (<||) p

p08pick

p08pick :: pick

Thus g can be any functional subset of:


                                     (k° >< k°) . cat° . k

Section 4

p09

p09 :: property

We show only the proof of the first membership, for which we show that:


  (p09)  ( ws ++ sort xs, ys ) ∈ iterF ( ws, zs ) <== xs djU ys = zs ∧ xs ≤ ys

p09proof

p09proof :: proof

Property (p09) can be proved by induction on the size of zs.

For zs = ∅ the property trivially holds. The nonempty case is shown below:


 ( ws ++ sort xs, ys ) ∈ iterF ( ws, zs )

≡      { zs /= ∅ , let z = min zs }

( ws ++ sort xs, ys ) = ( ws, zs ) ∨ ( ws ++ sort xs, ys ) ∈ iterF ( ws ++ [ z ] , zs − { z })

<==      { for non-empty xs, let x = min xs }

( xs = ∅ ∧ ys = zs ) ∨ ( ws ++ [ x ] ++ sort ( xs − { x }) , ys ) ∈ iterF ( ws ++ [ z ] , zs − { z })

<==      { induction }

( xs = ∅ ∧ ys = zs ) ∨ (( xs − { x } djU ys = zs − { z } ∧ xs ˙ ≤ ys ∧ x = z )

≡ xs djU ys = zs ∧ xs  ≤ ys.

p09pick

p09pick :: pick

By Lemma 6 we may thus pick:


   (p09pick)    g zs =  ( xs    , ys ) 

       for any           xs djU ys = zs ∧ xs ≤ ys. 

p10a

p10a :: property


 scanl (⊕) ι⊕ ( xs ++ ys ) 
      | xs' ++ [ x ] <- scanl (⊕) ι⊕ xs
      = xs' ++ [ x ] ++  map ( x ⊕) ( scanl (⊕) ι⊕ ys )

p10b

p10b :: property


 scan ( e, [ ])      = [ ]
 scan ( e, x ∶ xs ) = e ∶ scan ( e ⊕ x, xs )

p10c

p10c :: property


 scan ( e, xs ++ [ x ])  = scan ( e, xs ) ++ [ e ⊕ reduce xs ] ,
      where 
                reduce = hom (⊕) id ι⊕

p10d

p10d :: property


 scan = unfoldr (||>) p 
      = unfoldl (<||) p
      where               p = null . snd 

p10e

p10e :: property


                     (||>) ( e, x ∶ xs )     = ( e, ( e ⊕ x, xs ))
                     (<||) ( e, xs ++ [ ]) = (( e, xs ) , e ⊕ reduce xs )

where the domains of (||>) and (<||) are pairs whose second components are non-empty.

p10g

p10g :: property

To construct g, we show that for xs ++ ys = zs we have:


 ( scan ( e, xs ) ,      ( e ⊕ reduce xs, ys )) ∈ unfoldrp (||>)  p ( e, zs )
 (      ( e, xs ) , scan ( e ⊕ reduce xs, ys )) ∈ unfoldlp (<||)  p ( e, zs )

p10

p10 :: property


   (p10)   ( ws ++ scan ( e, xs ) , ( e ⊕ reduce xs, ys )) ∈ iterF ( ws, ( e, zs ))
             if xs ++ ys = zs.

p10proof

p10proof :: proof


    ( ws ++ scan ( e, xs ) , ( e ⊕ r xs, ys )) ∈ iterF ( ws, ( e, z ∶ zs ))
 ≡  ( ws ++ scan ( e, xs ) , ( e ⊕ r xs, ys )) = ( ws, ( e, z ∶ zs )) ∨
    ( ws ++ scan ( e, xs ) , ( e ⊕ r xs, ys )) ∈ iterF ( ws ++ [ e ] , ( e ⊕ z, zs ))

<==  { let xs = x ∶ xs' for the non-empty case }

( xs = [ ] ∧ ys = z ∶ zs ) ∨ ( ws ++ [ e ] ++ scan ( e ⊕ x, xs') , ( e ⊕ x ⊕ r xs', ys )) ∈ iterF ( ws ++ [ e ] , ( e ⊕ z, zs ))

<==      { induction }

( xs = [ ] ∧ ys = z ∶ zs ) ∨ ( xs' ++ ys = zs ∧ x = z )

≡ xs ++ ys = z ∶ zs.

p10pick

p10pick :: pick

Thus we pick:


      g ( e, zs )  =  (( e, xs ) , ( e ⊕ reduce xs, ys )) for some xs ++ ys = zs. 

p10h

p10h :: property


 build [ x ]         = L x
 build ( xs ++ ys )  |                               xs = [ ] ∧ ys = [ ]
                     = let t = build xs
                           u = build ys
                       in 
                          N t ( val t ⊕ val u ) u

p10i

p10i :: property

with the invariant that:


   val ( build xs ) = reduce xs:

p10j

p10j :: property

The key to construct an efficient implementation of scan is to perform build in a separate phase and use only the results stored in the tree.

That is, we wish to construct some function f such that:


 scan ( e, xs ) = f ( e, build xs )

The singleton case is easy:


 f ( e, L x ) = [ e ] . 

For inputs of length at least two, we calculate (for non-empty xs and ys):


 scan ( e, xs ++ ys )
 = scan ( e, xs ) ++ scan ( e ⊕ reduce xs, ys )
 =      { since val ( build xs ) = reduce xs }
 scan ( e, xs ) ++ scan ( e ⊕ val ( build xs ) , ys )
 =      { induction: scan ( e, xs ) = f ( e, build xs )  }
 f ( e, build xs ) ++ f ( e ⊕ val ( build xs ) , build ys )
 =      { let N t v u = build ( xs ++ ys )  }
 f ( e, t ) ++ f ( e ⊕ val t, u )
 =      { let f ( e, N t   u ) = f ( e, t ) ++ f ( e ⊕ val t, u )  }
 f ( e, build ( xs ++ ys )) .

We rename f to acc since it accumulates the result:


 scan ( e, [ ])      = [ ]
 scan ( e, xs )      = acc ( e, build xs )


 acc ( e, L x )      = [ e ]
 acc ( e, N t   u )  = acc ( e, t ) ++ acc ( e ⊕ val t, u )

Section 5

p11

p11 :: property


   (p11)   f . fill = U . ( id >< f ) ∧

p12

p12 :: property


   (p12)   f . fill = V . ( f′ >< id )

p13

p13 :: property


   (p13)   fill° . k = ( id >< k ) . mem . unfp↓ ( g↓, f↓, p )

By generating a tree “from the middle” we mean to find:


     g  :: b -> ( b, b )
     k′ :: b ->  Cxt a 

such that:


 k v  | p v = L ( f↓ v )
      | ( v↑, v↓ ) <- g v = fill ( k' v↑ ) ( k v↓ )

That is, the tree returned by k, if not a leaf, can be split into a context and a subtree that can be generated separately from the two seeds returned by g.

It suffices for g and k'to satisfy:


       fill° . k = (k >< k ) . g

Generating a tree “upwards” intuitively means to start from a leaf and find the path back to the root.

With application of Theorem 7 in mind, we want to come up with a function:


   unfp↑ ( g↑, f↑ ) :: b -> [( b, Tree a )]

that satisfies, for some k':

p14

p14 :: property


   (p14)   fill° . k = ( k' >< id ) . mem . unfp↑ ( g↑, f↑ )

p15

p15 :: property



(p15)   fill° . k   = ( k' >< k ) . ( k'° >< k° ) . fill° .k
         if        k   = unf↓ ( g↓, f↓, p ) = mem .  unf↑ ( g↑, f↑ )
                   k'  = mem . cxt↑ g↑ v. 

To compute k “from the middle”, we may pick g to be a subset of


 ( k'° >< k°) . fill° . k

Section 6