Pearl.GaDtTLHT.References

Description

References

Original Paper

Original References

r01 Reference 01

r01 :: a

R. S. Bird. An introduction to the theory of lists. In M. Broy, editor, Logic of Programming and Calculi of Discrete Design, volume 36 of NATO ASI Series F, pages 3-42. Springer-Verlag, New York, NY, USA, 1987. Technical Monograph PRG-56.

PDF: https://www.cs.ox.ac.uk/files/3378/PRG56.pdf

r02 Reference 02

r02 :: a

G. E. Blelloch. Prefix sums and their applications. In J. E. Reif, editor, Synthesis of Parallel Algorithms. Morgan Kaufmann Publishers Inc., San Francisco, CA, USA, 1993.

PDF: http://www.cs.cmu.edu/~guyb/papers/Ble93.pdf

r03 Reference 03

r03 :: a

M. I. Cole. Algorithmic Skeletons: Structural Management of Parallel Computation. MIT Press, 1989.

PS.GZ: http://www.dcs.ed.ac.uk/home/mic/skeletonbook.ps.gz

r04 Reference 04

r04 :: a

E. W. Dijkstra. A Discipline of Programming. Prentice Hall PTR, 1976.

Google Books: http://books.google.com/books/about/A_discipline_of_programming.html?id=MsUmAAAAMAAJ

r05 Reference 05

r05 :: a

A. Geser and S. Gorlatch. Parallelizing functional programs by generalization. Journal of Functional Programming, 9(6):649-673, 1999.

PDF: http://citeseerx.ist.psu.edu/viewdoc/download?doi=10.1.1.30.3544&rep=rep1&type=pdf

r06 Reference 06

r06 :: a

J. Gibbons. The third homomorphism theorem. Journal of Functional Programming, 6(4):657-665, 1996.

PDF: http://citeseerx.ist.psu.edu/viewdoc/download?doi=10.1.1.45.2247&rep=rep1&type=pdf

r07 Reference 07

r07 :: a

J. Gibbons, G. Hutton, and T. Altenkirch. When is a function a fold or an unfold? Electr. Notes Theor. Comput. Sci., 44(1), 2001.

PDF: http://www.cs.nott.ac.uk/~gmh/when.pdf

r08 Reference 08

r08 :: a

G. P. Huet. The zipper. Journal of Functional Programming, 7(5): 549-554, 1997.

PDF: http://yquem.inria.fr/~huet/PUBLIC/zip.pdf

r09 Reference 09

r09 :: a

A. Morihata, K. Matsuzaki, Z. Hu, and M. Takeichi. Program parallelization by candidate generation and conformity testing. IPSJ Transaction on Programming, 2(2):132-143, 2009. (In Japanese).

PDF: http://www.bookpark.ne.jp/cm/ipsj/search.asp?flag=6&keyword=IPSJ-TPRO0202011&mode=PDF

r10 Reference 10

r10 :: a

A. Morihata, K. Matsuzaki, Z. Hu, and M. Takeichi. The third homomorphism theorem on trees: Downward & upward lead to divide-and-conquer. In Proceedings of the 36th ACM SIGPLAN-SIGACT Symposium on Principles of Programming Languages, POPL 2009, Savannah, Georgia, USA, January 21-23, 2009, pages 177-185. ACM, 2009.

PDF: http://research.nii.ac.jp/~hu/pub/popl09.pdf

r11 Reference 11

r11 :: a

K. Morita, A. Morihata, K. Matsuzaki, Z. Hu, and M. Takeichi. Automatic inversion generates divide-and-conquer parallel programs. In J. Ferrante and K. S. McKinley, editors, Proceedings of the ACM SIGPLAN 2007 Conference on Programming Language Design and Implementation, San Diego, California, USA, June 10-13, 2007, pages 146-155. ACM, 2007.

PDF: http://takeichi.ipl-lab.org/pub/PLDI2007.pdf