Pearl.GaDtTLHT.Section02

Description

The Third-List Homomorphism Theorem

h :: property (h is a list homomorphism) denoted (hom f k e)

A function h :: [a] -> b is a [list homoprhism] if there exists:

 e :: b

 k :: a -> b

 f :: (b >< b) -> b

such that:


   h [] = e
   h [x] = k x
   h (xs ++ ys) = f (h xs, h ys)

In such a case, we denote h by hom f k e.

The equations imply that f is associative, on the range of h, with unit e.

To compute a list homorphism h, one may

split the list arbitrarily into two parts

recursively compute h on both parts

and combine the results using f

implying a potential for parallel computation. If f and k are constant-time operations, a list homorphism can be evaluted in time O( n / p + log p )

where

n is the length of the list, and

p is the number of processors

resulting in almost linear speedups with respect to p.

p01 :: [a1] -> property (e is a unit for a fold)

As is well known, in the world of sets and total functions, the equations:


   (p01)       h  [    ]  =   e
   (p01)       h  (x:xs)  =       (<||) (x, h xs)
   (p01a)      h          = foldr (<||) e

have (p01a) as unique solution for h :: [a] -> b.

p01a :: property foldr with (a leftwards function) and (a unit e) is unique


   (p01a)      h          = foldr (<||) e

e :: property (e is a unit for a fold)

e is the unit for the fold, ie, it is like the [] we replace at the end of a list.

Uncurrying some familiar functions

(<||) :: (a, b) -> b

(||>) :: (a, b) -> a

We deviate from the standard and let (<||) be uncurried since it is more convenient in point-free style, where programs are described by function composition rather than application.

cons :: (a, [a]) -> [a]

In fact, we will also introduce uncurried constructor:


   cons ( x, xs ) = x ∶ xs

a product function

(><) :: (l -> r) -> (u -> d) -> (l, u) -> (r, d)

p01b :: property p01 [pointfree style]

Thus (p01):


   (p01)       h  [    ]  =   e
   (p01)       h  (x:xs)  = (<||) (x, h xs)

can be written:


   (p01b)     h . cons = (<||) . (id >< h)

foldrr: a resumed foldr

foldrr :: ((a, b) -> b) -> ([a], b) -> b

We define a variation of foldr that takes the base case as an extra argument.

It can be seen as a resumed version of foldr (hence the suffix r in the name).

p02 :: property (foldrr is foldr [resumed])

That is, if:


   (p01a)      h          = foldr (<||) e

then one can show:


   (p02) h (xs ++ ys)   = foldrr (<||) ( xs , h ys )

by induction on xs.

cat :: ([a], [a]) -> [a]

p03 :: property p02 [pointfree style]

Let:

 cat (xs,ys) = xs ++ ys

then equation (p02):

(p02) h (xs ++ ys) = foldrr (<||) ( xs , h ys )

can be point-free as:

(p03) h . cat      = foldrr (<||) . ( id >< h )

DLists

unsnoc :: [a] -> ([a], a)

snoc :: ([a], a) -> [a]

Symmetrically, let:

 snoc (xs, x) = xs ++ [x]

p03b :: property (foldl with ||> (pronounced forward)) is the unique solution for snoclists

It is known that

 foldl (||>) e

is the unique solution for h :: [a] -> b

in:

p03h :: property p03b [pointfree]


   (p03h)  h [] = e
             h . snoc = (||>) . ( h >< id )

where (||>) :: (b,a) -> b.

foldlr: a resumed foldl

foldlr :: ((b, a) -> b) -> (b, [a]) -> b

Defining resumable foldl as:


   foldlr (||>) ( e   ,  []              ) = e
   foldlr (||>) ( e   , unsnoc -> (xs,x) ) = (||>) ( foldlr (||>) (e,xs)  ,  x  ) 

p04 :: property (dual to p03) [pointfree]

We have, if h = foldl (||>) e , that:

 (p04)  h . cat = foldlr (||>) . ( h >< id )

A List Homomorphism

Some Theorems

Theorem 1

t01 :: theorem (second list homomoprhism theorem)

Theorem:

(the 2nd list-homomorphism theorem [r01]).

If


        h = hom f k e

Then


        h = foldr <|| e
        h = foldl ||> e

Where


        (<||) (x, v) = f (k x,   v)
        (||>) (v, x) = f (  v, k x)

Proof

t01proof :: theorem proof [t01]

Proof.

A complete proof is given in [r01].

Theorem 2

t02 :: theorem (third list homomorphism theorem)

Somewhat surprisingly, if a function can be computed both by a foldr and a foldl, it is a list homomorphism.

Theorem:

(the 3rd list-homomorphism theorem [r06])


        h = foldr <|| e
        h = foldl ||> e

implies


        h = hom f k e

for some f and k.

Proof

t02proof :: theorem proof [t02]

Proof.

The only possible choice for k is:


        k x = h [x]

The aim is to find f such that:


        h . cat = f . (h >< h)

A function fI is called a right inverse of f if, for all y in the range of f, we have f (fI y) = y. Equivalently, f . fI . f = f.

In a set-theoretical model, a right inverse always exists but may not be unique.

While a semantical proof was given by [Gibbons r06], we will provide a proof having a much more equational flavour.

We reason:


                        h . cat
 { use (3) }           =
                        foldrr <|| . (id >< h)
 { h = h . hI . h
  and product functor} =
                        foldrr <|| . (id >< h)   . (id >< hI) . (id >< h)
 { (3) backwards
  and (4) forwards }   =
                        foldlr ||> . (h  >< id)  . (id >< hI) . (id >< h)
 { h = h . hI . h
  and product functor} =
                        foldlr ||> . (h  >< id)  . (hI >< hI) . (h >< h)
 { (4) backwards }     =
                         h . cat   . (hI >< hI)  . ( h >< h )

Comments

t02comments :: theorem comments [t02]

Comments.

Theorem 2 in fact provides hints how to construct list homomorphisms.

For example, since sum = foldr (+) 0 = foldl (+) 0, Theorem 2 states that sum can be written as:


      sum = hom f k 0 
        where k x        = sum [ x ] = x 
              f ( v, w ) = sum ( g v ++ g w ) 

for any right inverse g of sum.

One may simply pick


      g x = [x]

and f (v,w) simplifies to v + w.

Readers might have noticed something odd in the proof: the property much talked about, that h is both a foldr and a foldl, could be weakened - properties (p03) and (p04) were merely used to push h to the right. In fact, h . cat is never expanded in the proof.

One thus wonders whether there is something more general waiting to be discovered, which is indeed what we will see in the following sections. The syntactical approach makes such generalisations much easier to spot.

p04pick :: pick

Thus the theorem holds if we pick


                    f  = h . cat   . (hI >< hI)