Pearl.GaDtTLHT.Section03

Description

Unfolds! Left and Right!

The function unfoldr, a dual of foldr that generates a list from left to right, may be defined as follows: [note01]


 unfoldr (||>)  p v
              | p v                  = []
              | (x, v') <- (||>) v   = x : unfoldr '(||>') p v'

unfoldr :: (b -> (a, b)) -> (b -> Bool) -> b -> [a]

Symmetrically, the function unfoldl is defined by:


 unfoldl (<||)  p    v
              | p    v               = []
              | (v', x) <- (<||) v   = unfoldl (<||) p v' ++ [x]

unfoldl :: (b -> (b, a)) -> (b -> Bool) -> b -> [a]

Unfolds are (usually) coinductive...

Typically, unfolds are defined for coinductive, possibly infinite lists. Since we want the unfolded lists to have both a left end and a right end, and for another important technical reason to be mentioned later, our unfolds in this pearl return inductive, finite lists and require separate proofs that all successive applications of <|| and ||> eventually reach some v for which p v is True. Due to space constraints, however, the proof of termination is usually treated informally.

A definition for unhom

Finally, we denote a function k ∶∶ b → [ a ] by unhom g f p q it it satisfies:


 unhom g f p q = k
      where 
      k v | p v          = []
          | q v          = [f v]
          | (l,r) <- g v = k l ++ k r

unhom :: (t -> (t, t)) -> (t -> a) -> (t -> Bool) -> (t -> Bool) -> t -> [a]

We also demand that successive applications of g eventually produce seeds for which either p or q is true, and thus k generates finite lists.

Regrettably, Theorem 1 does not dualise to unfolds.

Lemma 3

t03 :: theorem

Lemma.

There exist h such that:

 h = unhom g f p q

for which there is no corresponding:


        h = unfoldr (||>) p
          = unfoldl (<||) p

Proof: by Example

t03proof :: theorem proof (by example)

Proof:

Let example0 :: Int -> [Int] generate a list of 1s of length 2^n for a given n.

It can be defined:


   example0 :: Int -> [Int]
   example0 = unhom
                  (\n -> (n - 1, n - 1))   {-g-}
                  (\0 -> 1)                {-f-}
                  (0 >)                    {-p-}
                  (0 ==)                   {-q-}

But it can not be defined by an unfoldl or unfoldr. An intuitive reason is that there is not always an m such that 2^n - 1 = 2^m. One may use theories of Gibbons et al. [r07] for proof.

Example

example0 :: Int -> [Int]

>>> map example0 [1..4]
[[1,1],[1,1,1,1],[1,1,1,1,1,1,1,1],[1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1]]

>>> map (length . example0) [1..10]
[2,4,8,16,32,64,128,256,512,1024]

Theorem 02 for Unfolds?

question :: a

Does theorem 2 have an unfold counterpart?

To establish that:


   k = unhom g f p' q

Given:


   k = unfoldr (||>) p
     = unfoldl (<||) p

We may simply pick:


 k = unhom g  f   p'  q
              |   |   |
              f   |   |      =     fst . (||>)        -- fixme?
                  p'  |      = p
                      q      = p . snd . (||>)        -- fixme?

The challenge is to construct g. To do so, we introduce some concepts dual to those for folds.

While foldrr and foldlr resume a partially computed fold, their duals unfoldrp and unfoldlp instead pause during the generation of lists.

During the generation of a list, there are many places where we may pause. The following functions unfoldrp and unfoldlp (p for pause) return, in a list, all intermediate lists and seeds during list generation:


 unfoldrp (||>) p v = (|>|) ([], v)
                where (|>|) (xs, v) | p v                = [(xs, v)]
                                    | (x, v') <- (||>) v = (xs, v) : (|>|) (xs ++ [x], v')


 unfoldlp (<||) p v = (|<|) (v, []) 
                where (|<|) (v, xs) | p v                 = [(v,xs)]
                                    | (v', x) <- (<||) v  = (|<|) (v', x:xs) ++ [(v,xs)]

unfoldrp :: (b -> (a, b)) -> (b -> Bool) -> b -> [([a], b)]

unfoldlp :: (b -> (b, a)) -> (b -> Bool) -> b -> [(b, [a])]

Traces of Unfolding

We may think of unfoldrp and unfoldlp as returning the traces of unfolding.

For example:


   unfoldrp (||>)  p  v0

Yields the list:


   [  ([], v0) , ( [x1], v1 ) , ( [x1,x2] , v2 ), ( [x1,x2,x3], v3 ) ... ]

If:


    ( x_i+1 , v_i+1 ) = (||>) v_i

A crucial property

A crucial property relating unfoldr and unfoldrp is that if:


   k = unfoldr (||>)  p
   k = unfoldl (<||)  p

we have both:


   (p05)       splits . k = map (id >< k) . unfoldrp (||>)
   (p06)       splits . k = map (k >< id) . unfoldlp (<||)

where


   splits :: [a] -> ( [a], [a] )

Notice how they resemble the converses of (p03) and (p04):


   (p03)       h . cat = foldrr (<||) . ( id >< h )
   (p04)       h . cat = foldlr (||>) . ( h >< id )


   (p05)       splits . k = map (id >< k) . unfoldrp (||>)
   (p06)       splits . k = map (k >< id) . unfoldlp (<||)

the functions next to the composition (.) are swapped

instead of f . cat we have splits . k on the left hand sides

(id >< k) in (5) is on the left-hand side of (.), and is lifted to lists by map due to the type, etc.

p05 :: property


   (p05)       splits . k = map (id >< k) . unfoldrp (||>)

p06 :: property


   (p06)       splits . k = map (k >< id) . unfoldlp (<||)

splits :: [a] -> ([a], [a])

>>> splits [1]
[ ([] , [1]) , ( [1], [] ) ]

>>> splits [1,2]
[ ([], [1,2]), ([1], [2]), ([1,2], []) ]

A Relational Perspective

We could have proceeded from (p05) and (p06) to construct g and thereby prove the dual theorem. However, we would like to take a bold approach and ask the readers to see paused unfolding as a relation.

Functional progammers appear to regard relations as an arcane creation, which is an unfortunate misunderstanding.

Dijkstra argued that, for program derivation, non-determinism should be the norm and determinism a special case [r04], and this pearl is indeed a case where one has to bring up relations to appreciate the nice duality between the theorems for folding and unfolding.

Relations: Introduction

For this pearl, we need only a one-paragraph introduction to relations.

To say that a function f:


  f :: a -> b    "is a subset of"    (b >< a)
                      where:         ( y      , x )  ∈   f
  means that:
       y  is a result to which
       f  maps 
       x  (the argument) 
  with the constraint: 
                                     ( y      , x )  ∈   f
     and                             ( |   y' , x )  ∈   f
                                       |   |
     which normally imply:             y = y'

This constraint is relaxed for relations, allowing x to be mapped to multiple ys.

Composition

Composition of relations is defined by:


   (z, x) ∈ (R . S) == ( (exists y :: (z,y) ∈ R) /\ ((y,x) ∈ S) )

Converse

Given:

 R  :: a -> b

Its converse:

 R° :: b -> a

is defined by:


  (x,y) ∈ R° == (y,x) ∈ R

Laws

We therefore have:


            (  R°  )°   = R          -- ( ° on (R°) is the original R )
  And:         id°      = id         -- ( ° on  id  is id )
  Also:     (R >< S)°   = (R° >< S°)   -- ( ° distributes covariantly      over (><) )
  And:      (R  . S)°   = S°  . R°   -- ( ° distributes contravariantly  over (.)  )
                                               

In this pearl we view unfolds as the relational converse of folds, which is another reason why we restrict ourselves to finite lists -- our unfolds are actually converses of folds in disguise, and an unhom is the relational converse of a list homomoprhism.

Let the relation mem relate a list to any of its members:

 mem :: [a] -> a

and let:


        unfr (||>) p = mem . unfoldrp (||>) p
        unfl (<||) p = mem . unfoldlp (<||) p

While unfoldrp and unfoldlp generate the entire trace, unfr and unfl map the input seed to one arbitrary intermediate state.

That is:


  unfr (||>) p
       maps v0 to each of:
                           [ ( []     , v0  ) 
                           , ( [x1]   , v1  ) 
                           , ( [x1,x2], v2  )
                             ...
                           ]
  if:                        ( x_i+1  ,v_i+1) = (||>) v_i

p07 :: property

The equality (p05) can be expressed relationally.

If k = unfoldr (||>) p we have:


  (p07)     cat° . k = (id >< k ) . unfr (||>) p

Both sides relate the input seed, say v, to a pair of lists. On the left hand side, the list returned by k v is arbitrarily split into (xs,ys).

The equation says that xs ++ ys = k v if and only if xs can be generated by unfr (||>) p and ys can be generated by k starting from where unfr (||>) p left off.

p08 :: property

For k = unfoldl (<||) p, a symmetric property holds:


  (p08)     cat° . k = (k  >< id) . unfl (<||) p

Recall that our aim is to construct g. It suffices to find a g such that:

 cat° . k = (k >< k) . g

That is:

 
  k v     = xs ++ ys   iff:
            ||    ||   
            xs    ||   xs and:
                  ys   ys       can be generated 
                                from seeds produced by g v.

Notice that by applying the converse operator (_)* to both sides of (p07) and (p08), we get equations that are almost (p03) and (p04) apart from having converses of k and unfolds in the formulae:


       k° . cat = (unfr (||>) p)° . (id >< k°)
       k° . cat = (unfl (<||) p)° . (k° >< id)

Theorem 4

t04 :: theorem t02 [on relations]

The proof of Theorem 2, however, proceeds the same way even if the components are not functions! In the realm of relations, part of the proof of Theorem 2 can be generalized to:

Theorem.


 If:
       R . R° . R     =    R                   (R has converse)

Then:
                    R . cat = Sfwd     . (id >< R )
                            = Srev     . (R  >< id)
 { ... }                    = R  . cat . (R° >< R°) . (R >< R)

Proof

t04proof :: theorem proof deferred

Proof.

The same as that of Theorem 2. We will prove a more general Theorem 7 later.

Corollary 5 <- Theorem 4

t05 :: theorem (corollary to t04)

The desired dual theorem thus follows:

Theorem.


 If:
   k  =  unhom      g  f  p  q
                    |  |  |  |
                    ?  ?  ?  ?  -- ^ for some g , f , p , and q
 Then:                    |  
   k  =  unfoldr (||>)    p
      =  unfoldl (<||)    p

Proof

t05proof :: proof (using p07) (using p08) [th04]

Proof.

We have already talked about f and q, and now we aim to find g such that:


        cat° . k = ( k >< k ) . g

Using properties (p07) and (p08) as antecedents of Theorem 4 we get:


 With:
      (p07)    cat° . k = (id >< k ) . unfr  (||>)  p
      (p08)    cat° . k = (k  >< id) . unfl  (<||)  p

We get:
               cat° . k = (k  >< k ) . (k° >< k°) . cat° . k

Calculating g

The expression:

 (k° >< k°) . cat° . k

is not easy to simplify. To calculate g we often use Lemma 6 to to refine:

 (k° >< k°) . cat° . k

to:


       ((k° >< id) . unfr (||>) p)
   ∩   ((id >< k°) . unfl (<||) p)

For readers who are not familar with intersection and converses, Figure 1 offers some intuition.

Given a seed v, and both:


   unfr (||>) p
   unfl (<||) p

which non-deterministically pause somewhere and respectively generate (xs,v2) and (v1,ys). The intersection means that they stop when they "meet in the middle", that is, when


   k v1 = xs 
   k v2 = ys

The new pair of seeds, a possible result of g, is (v1,v2).

p08pick :: pick

Thus g can be any functional subset of:


                                     (k° >< k°) . cat° . k

Figure 1


                   ------\
                   || ||  \ 
                   || ||   \
                  (xs,v2)   \
                  /^         \ unfr (||>) p
               k /            \ 
                /              \
         (v1,v2)               (v)
                \              /
               k \            /
                  \v         / unfl (<||) p 
                  (v1,ys)   /
                   || ||   /
                   || ||  /
                   ------/
                        

Lemma 6

t06 :: theorem lemma

Lemma.

This lemma has a functional formulation that may be more friendly to readers. Writing the list membership predicate elem :: a -> [a] -> Bool in the Haskell Standard Prelude in infix position as (∈), we have:

Assume that:


   k  =  unfoldr (||>)    p
      =  unfoldl (<||)    p

A function g is a subset of:


                                     (k° >< k°) . cat° . k

If, for all v:


       g v = (  v1,   v2)
 Where:
             (k v1,   v2) ∈ unfoldrp (||>) p v
             (  v1, k v2) ∈ unfoldlp (<||) p v

From now on we will use Lemma 6 and restrict ourselves to functions when we calculate g.

t06proof :: proof

Proof:

Proof omitted!

Extra Stuff

(<||) :: (b, t) -> t

(||>) :: (t1, b) -> t

(|<|) :: a

(|>|) :: a

p :: a

f :: a

g :: a

q :: a

k :: a

unfl :: a

unfr :: (Eq a1, Eq a) => (a -> (a1, a)) -> (a -> Bool) -> a -> [[([a1], a)]] -> Bool

(/\) :: (a -> b) -> (a -> c) -> a -> (b, c)