Pearl.GaDtTLHT.Section05

Description

Documentation

t07 :: theorem [on relations]

Theorem.

As mentioned earlier, when looking at the proof of Theorem 2 in detail, one notices that further generalization is possible. Indeed, the presence of R . cat in Theorem 4 is superficial, and the theorem can still be generalized:


 R   = U . (id >< S )    -- (RUS)
     = V . (T  >< id)    -- (RVT)

where


     S . S° . S = S        -- (S invertable)
     T . T° . T = T        -- (T invertable)

imply


 R   = R . (T° >< S°) . (T >< S)

t07proof :: theorem proof [t07] [on relations]

Proof:


                     R = U . (id x S)
  { S invertable }     = U . (id x S) . (id x S°) . (id x S)
  { RUS, in reverse }  = R            . (id x S°) . (id x S)
  { RVT }              = V . (T x id) . (id x S°) . (id x S)
  { T° T = id ? }      = V . (T x id) . (T° x S°) . (T x S)
  { RVT, in reverse }  = R            . (T° x S°) . (T x S)

{ done }           R = R . (T° x S°) . (T x S)

t07comments :: theorem comments [t07]

Comments:

However, the conclusion of the theorem, that:

R = R . (T° x S°) . (T x S)

does not have much structure hinting at how this theorem can be useful.

The use cases we have found are when S is a sub-expression of R (for example, to prove [t05], we used:

 
  R     = k° . cat
  S = T = k°

We do not require S = T in general.

We use Theorem 7 to establish recursive equations about S, hoping to construct a terminating definition of S.

p11 :: property


   (p11)   f . fill = U . ( id >< f ) ∧

p12 :: property


   (p12)   f . fill = V . ( f′ >< id )

test55 :: a

p13 :: property


   (p13)   fill° . k = ( id >< k ) . mem . unfp↓ ( g↓, f↓, p )

By generating a tree “from the middle” we mean to find:


     g  :: b -> ( b, b )
     k′ :: b ->  Cxt a 

such that:


 k v  | p v = L ( f↓ v )
      | ( v↑, v↓ ) <- g v = fill ( k' v↑ ) ( k v↓ )

That is, the tree returned by k, if not a leaf, can be split into a context and a subtree that can be generated separately from the two seeds returned by g.

It suffices for g and k'to satisfy:


       fill° . k = (k >< k ) . g

Generating a tree “upwards” intuitively means to start from a leaf and find the path back to the root.

With application of Theorem 7 in mind, we want to come up with a function:


   unfp↑ ( g↑, f↑ ) :: b -> [( b, Tree a )]

that satisfies, for some k':

p14 :: property


   (p14)   fill° . k = ( k' >< id ) . mem . unfp↑ ( g↑, f↑ )

the current tree as the left child

v r the seed for the right child

and v' the seed going up.

p15 :: property



(p15)   fill° . k   = ( k' >< k ) . ( k'° >< k° ) . fill° .k
         if        k   = unf↓ ( g↓, f↓, p ) = mem .  unf↑ ( g↑, f↑ )
                   k'  = mem . cxt↑ g↑ v. 

To compute k “from the middle”, we may pick g to be a subset of


 ( k'° >< k°) . fill° . k