Pearl.GaDtTLHT.Theorems

Description

All of the theorems are collected here, in one place, for convenience.

Section 2

t01 Second LHT

t01 :: theorem (second list homomoprhism theorem)

Theorem:

(the 2nd list-homomorphism theorem [r01]).

If


        h = hom f k e

Then


        h = foldr <|| e
        h = foldl ||> e

Where


        (<||) (x, v) = f (k x,   v)
        (||>) (v, x) = f (  v, k x)

Proof

t01proof :: theorem proof [t01]

Proof.

A complete proof is given in [r01].

t02 Third LHT

t02 :: theorem (third list homomorphism theorem)

Somewhat surprisingly, if a function can be computed both by a foldr and a foldl, it is a list homomorphism.

Theorem:

(the 3rd list-homomorphism theorem [r06])


        h = foldr <|| e
        h = foldl ||> e

implies


        h = hom f k e

for some f and k.

Proof

t02proof :: theorem proof [t02]

Proof.

The only possible choice for k is:


        k x = h [x]

The aim is to find f such that:


        h . cat = f . (h >< h)

A function fI is called a right inverse of f if, for all y in the range of f, we have f (fI y) = y. Equivalently, f . fI . f = f.

In a set-theoretical model, a right inverse always exists but may not be unique.

While a semantical proof was given by [Gibbons r06], we will provide a proof having a much more equational flavour.

We reason:


                        h . cat
 { use (3) }           =
                        foldrr <|| . (id >< h)
 { h = h . hI . h
  and product functor} =
                        foldrr <|| . (id >< h)   . (id >< hI) . (id >< h)
 { (3) backwards
  and (4) forwards }   =
                        foldlr ||> . (h  >< id)  . (id >< hI) . (id >< h)
 { h = h . hI . h
  and product functor} =
                        foldlr ||> . (h  >< id)  . (hI >< hI) . (h >< h)
 { (4) backwards }     =
                         h . cat   . (hI >< hI)  . ( h >< h )

Comments

t02comments :: theorem comments [t02]

Comments.

Theorem 2 in fact provides hints how to construct list homomorphisms.

For example, since sum = foldr (+) 0 = foldl (+) 0, Theorem 2 states that sum can be written as:


      sum = hom f k 0 
        where k x        = sum [ x ] = x 
              f ( v, w ) = sum ( g v ++ g w ) 

for any right inverse g of sum.

One may simply pick


      g x = [x]

and f (v,w) simplifies to v + w.

Readers might have noticed something odd in the proof: the property much talked about, that h is both a foldr and a foldl, could be weakened - properties (p03) and (p04) were merely used to push h to the right. In fact, h . cat is never expanded in the proof.

One thus wonders whether there is something more general waiting to be discovered, which is indeed what we will see in the following sections. The syntactical approach makes such generalisations much easier to spot.

Section 3

t03 Lemma 3

t03 :: theorem

Lemma.

There exist h such that:

 h = unhom g f p q

for which there is no corresponding:


        h = unfoldr (||>) p
          = unfoldl (<||) p

Proof

t03proof :: theorem proof (by example)

Proof:

Let example0 :: Int -> [Int] generate a list of 1s of length 2^n for a given n.

It can be defined:


   example0 :: Int -> [Int]
   example0 = unhom
                  (\n -> (n - 1, n - 1))   {-g-}
                  (\0 -> 1)                {-f-}
                  (0 >)                    {-p-}
                  (0 ==)                   {-q-}

But it can not be defined by an unfoldl or unfoldr. An intuitive reason is that there is not always an m such that 2^n - 1 = 2^m. One may use theories of Gibbons et al. [r07] for proof.

t04 Third LHT (on Relations)

t04 :: theorem t02 [on relations]

The proof of Theorem 2, however, proceeds the same way even if the components are not functions! In the realm of relations, part of the proof of Theorem 2 can be generalized to:

Theorem.


 If:
       R . R° . R     =    R                   (R has converse)

Then:
                    R . cat = Sfwd     . (id >< R )
                            = Srev     . (R  >< id)
 { ... }                    = R  . cat . (R° >< R°) . (R >< R)

Proof

t04proof :: theorem proof deferred

Proof.

The same as that of Theorem 2. We will prove a more general Theorem 7 later.

t05 Dual to Third LHT

t05 :: theorem (corollary to t04)

The desired dual theorem thus follows:

Theorem.


 If:
   k  =  unhom      g  f  p  q
                    |  |  |  |
                    ?  ?  ?  ?  -- ^ for some g , f , p , and q
 Then:                    |  
   k  =  unfoldr (||>)    p
      =  unfoldl (<||)    p

Proof

t05proof :: proof (using p07) (using p08) [th04]

Proof.

We have already talked about f and q, and now we aim to find g such that:


        cat° . k = ( k >< k ) . g

Using properties (p07) and (p08) as antecedents of Theorem 4 we get:


 With:
      (p07)    cat° . k = (id >< k ) . unfr  (||>)  p
      (p08)    cat° . k = (k  >< id) . unfl  (<||)  p

We get:
               cat° . k = (k  >< k ) . (k° >< k°) . cat° . k

t06 Lemma 6

t06 :: theorem lemma

Lemma.

This lemma has a functional formulation that may be more friendly to readers. Writing the list membership predicate elem :: a -> [a] -> Bool in the Haskell Standard Prelude in infix position as (∈), we have:

Assume that:


   k  =  unfoldr (||>)    p
      =  unfoldl (<||)    p

A function g is a subset of:


                                     (k° >< k°) . cat° . k

If, for all v:


       g v = (  v1,   v2)
 Where:
             (k v1,   v2) ∈ unfoldrp (||>) p v
             (  v1, k v2) ∈ unfoldlp (<||) p v

From now on we will use Lemma 6 and restrict ourselves to functions when we calculate g.

Proof

t06proof :: proof

Proof:

Proof omitted!

Section 5

t07 Unnamed Theorem

t07 :: theorem [on relations]

Theorem.

As mentioned earlier, when looking at the proof of Theorem 2 in detail, one notices that further generalization is possible. Indeed, the presence of R . cat in Theorem 4 is superficial, and the theorem can still be generalized:


 R   = U . (id >< S )    -- (RUS)
     = V . (T  >< id)    -- (RVT)

where


     S . S° . S = S        -- (S invertable)
     T . T° . T = T        -- (T invertable)

imply


 R   = R . (T° >< S°) . (T >< S)

Proof

t07proof :: theorem proof [t07] [on relations]

Proof:


                     R = U . (id x S)
  { S invertable }     = U . (id x S) . (id x S°) . (id x S)
  { RUS, in reverse }  = R            . (id x S°) . (id x S)
  { RVT }              = V . (T x id) . (id x S°) . (id x S)
  { T° T = id ? }      = V . (T x id) . (T° x S°) . (T x S)
  { RVT, in reverse }  = R            . (T° x S°) . (T x S)

{ done }           R = R . (T° x S°) . (T x S)

Comments

t07comments :: theorem comments [t07]

Comments:

However, the conclusion of the theorem, that:

R = R . (T° x S°) . (T x S)

does not have much structure hinting at how this theorem can be useful.

The use cases we have found are when S is a sub-expression of R (for example, to prove [t05], we used:

 
  R     = k° . cat
  S = T = k°

We do not require S = T in general.

We use Theorem 7 to establish recursive equations about S, hoping to construct a terminating definition of S.